+++ src/main/scala/coulomb/package.scala
@@ -15,6 +15,7 @@ limitations under the License.
 */
 
 import spire.math._
+import coulomb.infra.ConvertableUnits
 
 package object coulomb {
   trait Unitless
